Elixir v1.8.0-dev Calendar behaviour

This module defines the responsibilities for working with calendars, dates, times and datetimes in Elixir.

Currently it defines types and the minimal implementation for a calendar behaviour in Elixir. The goal of the Calendar features in Elixir is to provide a base for interoperability instead of full-featured datetime API.

For the actual date, time and datetime structures, see Date, Time, NaiveDateTime and DateTime.

Note the year, month, day, etc. designations are overspecified (i.e. an integer instead of 1..12 for months) because different calendars may have a different number of days per month, months per year and so on.

Types

calendar()

calendar() :: module()

A calendar implementation

date()

date() :: %{
  optional(any()) => any(),
  :calendar => calendar(),
  :year => year(),
  :month => month(),
  :day => day()
}

Any map/struct that contains the date fields

day()

day() :: integer()

day_fraction()

day_fraction() ::
  {parts_in_day :: non_neg_integer(), parts_per_day :: pos_integer()}

The internal time format is used when converting between calendars.

It represents time as a fraction of a day (starting from midnight). parts_in_day specifies how much of the day is already passed, while parts_per_day signifies how many parts there fit in a day.

hour()

hour() :: integer()

iso_days()

iso_days() :: {days :: integer(), day_fraction()}

The internal date format that is used when converting between calendars.

This is the number of days including the fractional part that has passed of the last day since 0000-01-01+00:00T00:00.000000 in ISO 8601 notation (also known as midnight 1 January BC 1 of the proleptic Gregorian calendar).

microsecond()

microsecond() :: {0..999_999, 0..6}

Microseconds with stored precision.

The precision represents the number of digits that must be used when representing the microseconds to external format. If the precision is 0, it means microseconds must be skipped.

minute()

minute() :: integer()

month()

month() :: integer()

naive_datetime()

naive_datetime() :: %{
  optional(any()) => any(),
  :calendar => calendar(),
  :year => year(),
  :month => month(),
  :day => day(),
  :hour => hour(),
  :minute => minute(),
  :second => second(),
  :microsecond => microsecond()
}

Any map/struct that contains the naive_datetime fields

second()

second() :: integer()

std_offset()

std_offset() :: integer()

The time zone standard offset in seconds (not zero in summer times)

time()

time() :: %{
  optional(any()) => any(),
  :hour => hour(),
  :minute => minute(),
  :second => second(),
  :microsecond => microsecond()
}

Any map/struct that contains the time fields

time_zone()

time_zone() :: String.t()

The time zone ID according to the IANA tz database (e.g. Europe/Zurich)

utc_offset()

utc_offset() :: integer()

The time zone UTC offset in seconds

year()

year() :: integer()

zone_abbr()

zone_abbr() :: String.t()

The time zone abbreviation (e.g. CET or CEST or BST etc.)

Functions

compatible_calendars?(calendar, calendar) (since 1.5.0)

compatible_calendars?(Calendar.calendar(), Calendar.calendar()) :: boolean()

Returns true if two calendars have the same moment of starting a new day, false otherwise.

If two calendars are not compatible, we can only convert datetimes and times between them. If they are compatible, this means that we can also convert dates as well as naive datetimes between them.

truncate(microsecond_tuple, atom) (since 1.6.0)

truncate(Calendar.microsecond(), :microsecond | :millisecond | :second) ::
  Calendar.microsecond()

Returns a microsecond tuple truncated to a given precision (:microsecond, :millisecond or :second).

Callbacks

date_to_string(year, month, day)

date_to_string(year(), month(), day()) :: String.t()

Converts the date into a string according to the calendar.

datetime_to_string(year, month, day, hour, minute, second, microsecond, time_zone, zone_abbr, utc_offset, std_offset)

datetime_to_string(
  year(),
  month(),
  day(),
  hour(),
  minute(),
  second(),
  microsecond(),
  time_zone(),
  zone_abbr(),
  utc_offset(),
  std_offset()
) :: String.t()

Converts the datetime (with time zone) into a string according to the calendar.

day_of_week(year, month, day)

day_of_week(year(), month(), day()) :: non_neg_integer()

Calculates the day of the week from the given year, month, and day.

day_rollover_relative_to_midnight_utc()

day_rollover_relative_to_midnight_utc() :: day_fraction()

Define the rollover moment for the given calendar.

This is the moment, in your calendar, when the current day ends and the next day starts.

The result of this function is used to check if two calendars rollover at the same time of day. If they do not, we can only convert datetimes and times between them. If they do, this means that we can also convert dates as well as naive datetimes between them.

This day fraction should be in its most simplified form possible, to make comparisons fast.

Examples

If, in your Calendar, a new day starts at midnight, return {0, 1}.

If, in your Calendar, a new day starts at sunrise, return {1, 4}.

If, in your Calendar, a new day starts at noon, return {1, 2}.

If, in your Calendar, a new day starts at sunset, return {3, 4}.

days_in_month(year, month)

days_in_month(year(), month()) :: day()

Returns how many days there are in the given year-month.

leap_year?(year)

leap_year?(year()) :: boolean()

Returns true if the given year is a leap year.

A leap year is a year of a longer length than normal. The exact meaning is up to the calendar. A calendar must return false if it does not support the concept of leap years.

months_in_year(year)

months_in_year(year()) :: month()

Returns how many months there are in the given year.

naive_datetime_from_iso_days(iso_days)

naive_datetime_from_iso_days(iso_days()) ::
  {year(), month(), day(), hour(), minute(), second(), microsecond()}

Converts iso_days/0 to the Calendar’s datetime format.

naive_datetime_to_iso_days(year, month, day, hour, minute, second, microsecond)

naive_datetime_to_iso_days(
  year(),
  month(),
  day(),
  hour(),
  minute(),
  second(),
  microsecond()
) :: iso_days()

Converts the given datetime (without time zone) into the iso_days/0 format.

naive_datetime_to_string(year, month, day, hour, minute, second, microsecond)

naive_datetime_to_string(
  year(),
  month(),
  day(),
  hour(),
  minute(),
  second(),
  microsecond()
) :: String.t()

Converts the datetime (without time zone) into a string according to the calendar.

time_from_day_fraction(day_fraction)

time_from_day_fraction(day_fraction()) ::
  {hour(), minute(), second(), microsecond()}

Converts day_fraction/0 to the Calendar’s time format.

time_to_day_fraction(hour, minute, second, microsecond)

time_to_day_fraction(hour(), minute(), second(), microsecond()) ::
  day_fraction()

Converts the given time to the day_fraction/0 format.

time_to_string(hour, minute, second, microsecond)

time_to_string(hour(), minute(), second(), microsecond()) :: String.t()

Converts the time into a string according to the calendar.

valid_date?(year, month, day)

valid_date?(year(), month(), day()) :: boolean()

Should return true if the given date describes a proper date in the calendar.

valid_time?(hour, minute, second, microsecond)

valid_time?(hour(), minute(), second(), microsecond()) :: boolean()

Should return true if the given time describes a proper time in the calendar.