Elixir v1.8.0-dev List.Chars protocol

The List.Chars protocol is responsible for converting a structure to a charlist (only if applicable).

The only function required to be implemented is to_charlist/1 which does the conversion.

The to_charlist/1 function automatically imported by Kernel invokes this protocol.

Types

t()

t() :: term()

Functions

to_charlist(term)

Converts term to a charlist.